Major refactor: splits resctrl tests, replace task_name with container_name and get_tasks with get_tids
